% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttr.calc.R
\name{ttr.calc}
\alias{ttr.calc}
\title{Converts INR Values to TTR}
\usage{
ttr.calc(day, ttr_interval, inr_min, inr_max,
patient_id, data, inr_variable, time_variable, id_variable,
print_number_of_inr_values = FALSE, print_matrix1 = FALSE,
print_matrix2 = FALSE)
}
\arguments{
\item{day}{The time of the desired TTR interval. TTR will be calculated for the period before the specified day.}

\item{ttr_interval}{The lenght of the desired TTR interval.}

\item{inr_min}{The lower limit of the INR target.}

\item{inr_max}{The upper limit of the INR target.}

\item{patient_id}{The individual patient number matching with the id_variable.}

\item{data}{The data frame including the INR values, patient numbers and time variable.}

\item{inr_variable}{The variable name in the data frame including the INR values.}

\item{time_variable}{The variable name including the time of INR measurements.}

\item{id_variable}{The variable name including the individual patient numbers.}

\item{print_number_of_inr_values}{If TRUE, will print a matrix including the TTR value and the number of INR measurements during the examined interval. The default is FALSE.}

\item{print_matrix1}{If TRUE, will print a matrix including all the INR values for the specified patients in addition to time periods, days in therapeutic range for each time interval, etc. The default is FALSE.}

\item{print_matrix2}{If TRUE, will print the aforementioned matrix for the desired time interval in which the first and last time in therapeutic range values are recalculated for the TTR calculation. The default is FALSE.}
}
\description{
This function converts INR (International Normalized Ratio) values to TTR (Time in Therapeutic Range) for the desired time and interval. The TTR values are calculated via the Rosendaal method (1). Any time periods exceeding the specified TTR interval from the previous measured INR value return '999'.
}
\examples{
##Not run:
patient_number <- ('BIO1005')
time <- c(3, 12, 70)
INR <- c(2.2, 2.6, 3.9)
INR_data <- data.frame(patient_number, time, INR)
ttr.calc(day = 65, ttr_interval = 60, inr_min = 2, inr_max = 3,
patient_id = 'BIO1005', data = INR_data, inr_variable = 'INR',
time_variable = 'time', id_variable = 'patient_number',
print_number_of_inr_values = TRUE)
## End(**Not run**)
}
\references{
Rosendaal F, Cannegieter S, van der Meer F, et al. A method to determine the optimal intensity of oral anticoagulant therapy. Thromb Haemost 1993; 69(3): 236–239.
}
